students/d3310/Ван Цзиян/lab3/lab1/requirements.txt
fastapi==0.115.13
uvicorn==0.34.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.9.2
jinja2==3.1.5
httpx==0.28.1
sqlmodel==0.0.20
email-validator==2.3.0
celery==5.3.4
redis==5.0.1 